Create some midi file parser methods in synth engine
a class in synth engine that will be used instead of receiver, that uses files(currently placed in assets, later we will pick them up from sd card). We are trying to parse the midi code in this file and pass it to very same method that were being used by the midi receiver-onsend method. this method is incomplete- figuring out the part where we construct the byte data from string midi code